pyproject.toml
[project]
name = "nagents"
version = "0.2.11"
description = "A lightweight, dependency-free LLM agent framework with direct HTTP-based provider integration"
readme = "README.md"
requires-python = ">=3.11, <4.0"
classifiers = [ "Development Status :: 4 - Beta", "Intended Audience :: Developers", "License :: OSI Approved :: MIT License", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13", "Programming Language :: Python :: 3.14", "Topic :: Scientific/Engineering :: Artificial Intelligence", "Topic :: Software Development :: Libraries :: Python Modules", "Typing :: Typed",]
keywords = [ "llm", "agent", "ai", "openai", "anthropic", "gemini", "claude", "gpt", "streaming", "async",]
dependencies = [ "aiohttp>=3.9.0", "aiosqlite>=0.20.0",]
[[project.authors]]
name = "Abbas Jafari"
email = "git@public.abja.dev"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.license]
text = "MIT"

[project.optional-dependencies]
dev = [ "pytest>=8.0", "pytest-asyncio>=0.24.0", "pytest-cov>=6.0.0", "ruff>=0.9.0", "mypy>=1.14.0", "pre-commit>=4.0.0",]
docs = [ "mkdocs>=1.6.0", "mkdocs-material>=9.5.0", "mkdocstrings[python]>=0.27.0",]

[project.urls]
Homepage = "https://github.com/abi-jey/nagents"
Repository = "https://github.com/abi-jey/nagents"
Documentation = "https://abi-jey.github.io/nagents/"
Issues = "https://github.com/abi-jey/nagents/issues"

[tool.ruff]
line-length = 120
target-version = "py311"
src = [ "src", "tests",]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
mypy_path = "src"
exclude = [ "tests/",]
ignore_missing_imports = true

[tool.ruff.lint]
select = [ "E", "F", "I", "UP", "B", "SIM", "TCH", "RUF",]
fixable = [ "ALL",]
ignore = [ "E501",]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = [ "tests",]
pythonpath = [ "src",]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"

[tool.coverage.run]
source = [ "src/nagents",]
branch = true

[tool.coverage.report]
exclude_lines = [ "pragma: no cover", "def __repr__", "raise NotImplementedError", "if TYPE_CHECKING:", "if __name__ == .__main__.:",]

[tool.ruff.lint.isort]
force-single-line = true
case-sensitive = true
known-first-party = [ "nagents",]

[tool.hatch.build.targets.wheel]
packages = [ "src/nagents",]

[tool.hatch.build.targets.sdist]
include = [ "/src", "/tests", "/docs", "LICENSE", "README.md", "pyproject.toml",]

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.7.1"
mkdocstrings = "^1.0.3"
mkdocstrings-python = "^2.0.2"

[tool.poetry.group.dev.dependencies]
setuptools = "^82.0.0"
qwen-tts = "^0.1.1"
python-dotenv = "^1.2.1"
pre-commit = "^4.5.1"

[tool.poetry.group.dev.dependencies.huggingface-hub]
extras = [ "cli",]
version = "<1.0.0"
